Reload frequency limit
1. Limit dnsmasq reload period to at least 5 seconds.
2. Recovered some lost code from previous versions due to problems when
syncing with GitHub.

+++ main/ip6neigh-svc.sh
@@ -22,7 +22,7 @@
 . /usr/lib/ip6neigh/ip6addr_functions.sh
 
 #Program definitions
-readonly VERSION="1.4.0"
+readonly VERSION="1.4.1"
 readonly CONFIG_FILE="/etc/config/ip6neigh"
 readonly HOSTS_FILE="/tmp/hosts/ip6neigh"
 readonly CACHE_FILE="/tmp/ip6neigh.cache"
@@ -74,6 +74,7 @@ config_get_bool MANUF_NAMES config manuf_names 1
 config_get PROBE_EUI64 config probe_eui64 1
 config_get_bool PROBE_IID config probe_iid 1
 config_get_bool LOAD_STATIC config load_static 1
+config_get_bool PROBE_HOST config probe_host 1
 config_get FLUSH config flush 1
 config_get FW_SCRIPT config fw_script
 config_get LOG config log 0
@@ -89,14 +90,36 @@ if [ -z "$DOMAIN" ]; then
 fi
 if [ -z "$DOMAIN" ]; then DOMAIN="lan"; fi
 
+#Asks dnsmasq to reload the hosts file if the pending flag is set
+reload_hosts() {
+	#Check the pending flag
+	[ "$reload_pending" = 1 ] || return 0
+	
+	local now
+	local diff
+	
+	#Current time
+	now=$(date +%s)
+
+	#Difference from the last reload time in seconds
+	diff=$(($now - $reload_time))
+	
+	#Reloads if the difference is more than 5 seconds with reload pending
+	if [ "$diff" -ge 5 ]; then
+		reload_time="$now"
+		reload_pending=0
+		killall -1 dnsmasq
+	fi	
+}
+
 #Adds entry to hosts file
 add() {
 	local name="$1"
 	local addr="$2"
 	echo "$addr $name" >> "$HOSTS_FILE"
-	killall -1 dnsmasq
 
 	logmsg "Added host: $name $addr"
+	reload_pending=1
 	return 0
 }
 
@@ -109,6 +132,7 @@ remove() {
 	mv "$TEMP_FILE" "$HOSTS_FILE"
 
 	logmsg "Removed host: $addr"
+	reload_pending=1
 	return 0
 }
 
@@ -151,6 +175,7 @@ rename() {
 	mv "$TEMP_FILE" "$CACHE_FILE"
 
 	logmsg "Renamed host: $oldname to $newname"
+	reload_pending=1
 	return 0
 }
 
@@ -184,6 +209,7 @@ remove_tmp_label() {
 	mv "$TEMP_FILE" "$HOSTS_FILE"
 
 	logmsg "Removed the ${tmp_label} label from the addresses with IID: ${iid}"
+	reload_pending=1
 	return 0
 }
 
@@ -247,8 +273,8 @@ add_probe() {
 	local addr
 	compress_addr addr "$1"
 	
-	#Do not add if the address already exist in some hosts file.
-	grep -q "^$addr[ ,"$'\t'"]" /tmp/hosts/* && return 0
+	#Do not add if the address already exist in some hosts file and unique flag was set on call.
+	[ "$2" -gt 0 ] && grep -q "^$addr[ ,"$'\t'"]" /tmp/hosts/* && return 0
 	
 	#Adds to the list
 	probe_list="${probe_list} ${addr}"
@@ -256,7 +282,7 @@ add_probe() {
 	return 0
 }
 
-#Probe addresses related to the supplied base address and MAC.
+#Probe addresses that are related to the supplied base address and MAC.
 probe_addresses() {
 	local name="$1"
 	local baseaddr="$2"
@@ -265,6 +291,20 @@ probe_addresses() {
 
 	#Initializes probe list
 	probe_list=""
+	
+	#Check if is configured for probing all the addresses from the same host
+	if [ "$PROBE_HOST" -gt 0 ]; then
+		#Get the address list for this host.
+		local hlist
+		local haddr
+		hlist=$(grep -E " ${name}(\.|$)" "$HOSTS_FILE" | cut -d ' ' -f1)
+		local OIFS="$IFS"
+		unset IFS
+		for haddr in $hlist; do
+			[ "$haddr" != "$addr" ] && add_probe "$haddr"
+		done
+		IFS="$OIFS"
+	fi
 
 	#Check if is configured for probing addresses with the same IID
 	local base_iid=""
@@ -273,10 +313,10 @@ probe_addresses() {
 		addr_iid64 base_iid "$baseaddr"
 		
 		#Probe same IID for different scopes than this one.
-		if [ "$scope" != 0 ]; then add_probe "fe80::${base_iid}"; fi
-		if [ "$scope" != 1 ] && [ -n "$ula_prefix" ]; then add_probe "${ula_prefix}:${base_iid}"; fi
-		if [ "$scope" != 2 ] && [ -n "$wula_prefix" ]; then add_probe "${wula_prefix}:${base_iid}"; fi
-		if [ "$scope" != 3 ] && [ -n "$gua_prefix" ]; then add_probe "${gua_prefix}:${base_iid}"; fi
+		if [ "$scope" != 0 ]; then add_probe "fe80::${base_iid}" 1; fi
+		if [ "$scope" != 1 ] && [ -n "$ula_prefix" ]; then add_probe "${ula_prefix}:${base_iid}" 1; fi
+		if [ "$scope" != 2 ] && [ -n "$wula_prefix" ]; then add_probe "${wula_prefix}:${base_iid}" 1; fi
+		if [ "$scope" != 3 ] && [ -n "$gua_prefix" ]; then add_probe "${gua_prefix}:${base_iid}" 1; fi
 	fi
 
 	#Check if is configured for probing MAC-based addresses
@@ -287,15 +327,15 @@ probe_addresses() {
 
 		#Only add to list if EUI-64 IID is different from the one that has been just added.
 		if [ "$eui64_iid" != "$base_iid" ]; then
-			if [ "$PROBE_EUI64" = "1" ] && [ "$scope" != 0 ]; then add_probe "fe80::${eui64_iid}"; fi
+			if [ "$PROBE_EUI64" = "1" ] && [ "$scope" != 0 ]; then add_probe "fe80::${eui64_iid}" 1; fi
 			if [ "$PROBE_EUI64" = "1" ] || [ "$scope" = 1 ]; then
-				if [ -n "$ula_prefix" ]; then add_probe "${ula_prefix}:${eui64_iid}"; fi
+				if [ -n "$ula_prefix" ]; then add_probe "${ula_prefix}:${eui64_iid}" 1; fi
 			fi
 			if [ "$PROBE_EUI64" = "1" ] || [ "$scope" = 2 ]; then
-				if [ -n "$wula_prefix" ]; then add_probe "${wula_prefix}:${eui64_iid}"; fi
+				if [ -n "$wula_prefix" ]; then add_probe "${wula_prefix}:${eui64_iid}" 1; fi
 			fi
 			if [ "$PROBE_EUI64" = "1" ] || [ "$scope" = 3 ]; then
-				if [ -n "$gua_prefix" ]; then add_probe "${gua_prefix}:${eui64_iid}"; fi
+				if [ -n "$gua_prefix" ]; then add_probe "${gua_prefix}:${eui64_iid}" 1; fi
 			fi
 		fi
 	fi
@@ -811,6 +851,10 @@ main_service() {
 
 	#Startup message
 	logmsg "Starting ip6neigh main service v${VERSION} for physdev $LAN_DEV with domain $DOMAIN"
+	
+	#Var initialization
+	reload_time=0
+	reload_pending=1
 
 	#Gets the IPv6 addresses from the LAN device.
 	ll_cidr=$(ip -6 addr show "$LAN_DEV" scope link 2>/dev/null | grep -m 1 'inet6' | awk '{print $2}')
@@ -918,7 +962,7 @@ main_service() {
 	echo "#Discovered IPv6 neighbors" >> "$HOSTS_FILE"
 
 	#Send signal to dnsmasq to reload hosts files.
-	killall -1 dnsmasq
+	reload_hosts
 
 	#Pings "all nodes" multicast address with source addresses from various scopes to speedup discovery.
 	ping6 -q -W 1 -c 3 -s 0 -I "$LAN_DEV" ff02::1 >/dev/null 2>/dev/null
@@ -963,6 +1007,7 @@ main_service() {
 		while IFS= read -r line
 		do
 			process $line
+			reload_hosts
 		done
 	
 	logmsg "Terminating the main service"
@@ -993,7 +1038,9 @@ snooping_service() {
 			[ -n "$addr" ] || continue
 			
 			#Check if the address already exists in any hosts file
-			if grep -q "^$addr[ ,"$'\t'"]" /tmp/hosts/* ; then continue; fi
+			if [ "$PROBE_HOST" != 1 ] && grep -q "^$addr[ ,"$'\t'"]" /tmp/hosts/* ; then
+				continue
+			fi
 				
 			#Ping the address to trigger a NS message from the router
 			sleep 1
